refactor: revamp inference plugin as class - deprecate function registering
