+++ core/src/plugins/index.ts
@@ -1,1 +1,3 @@
 export { ConversationalPlugin } from './conversational'
+export { InferencePlugin } from './inference'
+export { ModelPlugin } from './model'
