--- plugins/inference-plugin/@types/global.d.ts
@@ -1,3 +1,0 @@
-declare const PLUGIN_NAME: string;
-declare const MODULE_PATH: string;
-declare const INFERENCE_URL: string;
