+++ plugins/inference-plugin/src/@types/global.d.ts
@@ -1,0 +1,2 @@
+declare const MODULE: string;
+declare const INFERENCE_URL: string;
