+++ plugins/inference-plugin/src/module.ts
