+++ plugins/inference-plugin/tsconfig.json
@@ -8,6 +8,8 @@
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": false,
-    "skipLibCheck": true
-  }
+    "skipLibCheck": true,
+    "rootDir": "./src"
+  },
+  "include": ["./src"]
 }
